A new page for new users and a few minor changes.
Adds a new page call UserList which opens up when a user is not following
anyone.

Adds a new component called ProfileGlance which shows the users names,
about and a follow button.
It is being used in Follow and UserList pages.

Changes profile related URLs to the backend beacause of the changes made
to the backend.

Finally some minor css changes to improve visuals.